Move public headers out of `details`
This formally exposes the public interface of lucenaBAL for the purpose
of clarifying what is public and what is not. Note that these changes
do not break ABI or change the API in any way.
