+++ include/lucenaBAL/details/lbalLibrarySetup.hpp
@@ -17,7 +17,7 @@
 
 
 //	lbal
-#include <lucenaBAL/details/lbalConfig.hpp>
+#include <lucenaBAL/lbalConfig.hpp>
 #include <lucenaBAL/details/lbalCompilerSetup.hpp>
 #include <lucenaBAL/details/lbalDetectStandardLibrary.hpp>
 #include <lucenaBAL/details/lbalKnownVersions.hpp>
