+++ include/lucenaBAL/details/libraries/lbalStandardLibraryLibCppPostInitialization.hpp
@@ -17,7 +17,7 @@
 
 
 //	lbal
-#include <lucenaBAL/details/lbalConfig.hpp>
+#include <lucenaBAL/lbalConfig.hpp>
 #include <lucenaBAL/details/lbalDetectStandardLibrary.hpp>
 #include <lucenaBAL/details/lbalKnownVersions.hpp>
